+++ src/fides/api/ops/api/v1/endpoints/user_endpoints.py
@@ -1,9 +1,9 @@
 import json
-import logging
 from typing import Optional
 
 import jose.exceptions
 from fastapi import Depends, HTTPException, Security
+from loguru import logger
 from sqlalchemy.orm import Session
 from starlette.status import (
     HTTP_200_OK,
@@ -37,7 +37,6 @@ from fides.lib.oauth.oauth_util import extract_payload
 from fides.lib.oauth.schemas.user import UserPasswordReset, UserResponse, UserUpdate
 
 CONFIG = get_config()
-logger = logging.getLogger(__name__)
 router = APIRouter(tags=["Users"], prefix=V1_URL_PREFIX)
 
 
@@ -78,7 +77,7 @@ def update_user(
         )
 
     user.update(db=db, data=data.dict())
-    logger.info("Updated user with id: '%s'.", user.id)
+    logger.info("Updated user with id: '{}'.", user.id)
     return user
 
 
@@ -110,7 +109,7 @@ def update_user_password(
 
     current_user.update_password(db=db, new_password=b64_str_to_str(data.new_password))
 
-    logger.info("Updated user with id: '%s'.", current_user.id)
+    logger.info("Updated user with id: '{}'.", current_user.id)
     return current_user